+++ express-zod-api/src/endpoint.ts
@@ -212,12 +212,11 @@ export class Endpoint<
     return this.#def.getOperationId?.(method);
   }
 
-  async #parseOutput(output: z.input<OUT>) {
+  async #parseOutput(output: z.input<OUT>, config: CommonConfig) {
     try {
-      return (await parseMaybeAsync(
-        this.#def.outputSchema,
-        output,
-      )) as FlatObject;
+      return (await (config.trySyncValidation
+        ? parseMaybeAsync(this.#def.outputSchema, output)
+        : this.#def.outputSchema.parseAsync(output))) as FlatObject;
     } catch (e) {
       throw e instanceof z.ZodError ? new OutputValidationError(e) : e;
     }
@@ -236,6 +235,7 @@ export class Endpoint<
     response: Response;
     logger: ActualLogger;
     ctx: Partial<CTX>;
+    config: CommonConfig;
   }) {
     for (const mw of this.#def.middlewares || []) {
       if (
@@ -257,15 +257,19 @@ export class Endpoint<
 
   async #parseAndRunHandler({
     input,
+    config,
     ...rest
   }: {
     input: Readonly<FlatObject>;
     ctx: CTX;
     logger: ActualLogger;
+    config: CommonConfig;
   }) {
     let finalInput: z.output<IN>; // final input types transformations for handler
     try {
-      finalInput = await parseMaybeAsync(this.#def.inputSchema, input);
+      finalInput = config.trySyncValidation
+        ? await parseMaybeAsync(this.#def.inputSchema, input)
+        : await this.#def.inputSchema.parseAsync(input);
     } catch (e) {
       throw e instanceof z.ZodError ? new InputValidationError(e) : e;
     }
@@ -307,6 +311,7 @@ export class Endpoint<
         response,
         logger,
         ctx,
+        config,
       });
       if (response.writableEnded) return;
       if (method === ("options" satisfies CORSMethod))
@@ -314,10 +319,11 @@ export class Endpoint<
       const output = await this.#parseAndRunHandler({
         input,
         logger,
+        config,
         ctx: ctx as CTX, // ensured the complete CTX by writableEnded condition and try-catch
       });
       if (response.writableEnded) return; // Endpoint closed the stream (304)
-      result = { output: await this.#parseOutput(output), error: null };
+      result = { output: await this.#parseOutput(output, config), error: null };
     } catch (e) {
       result = { output: null, error: ensureError(e) };
     }